init session coockie

+++ src/main/java/com/afulvio/booklify/dto/OrderDTO.java
@@ -2,6 +2,8 @@ package com.afulvio.booklify.dto;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 public class OrderDTO {
 
@@ -29,6 +31,6 @@ public class OrderDTO {
 
     private String note;
 
-    private double total;
+    private BigDecimal total;
 
 }
